+++ app/Application/UserDataSource/UserDataSource.php
@@ -4,7 +4,9 @@ namespace App\Application\UserDataSource;
 
 use App\Domain\User;
 
-Interface UserDataSource
+interface UserDataSource
 {
     public function findByEmail(string $email): User;
+
+    public function getUsersIdsList(): array;
 }
